Added tests to reviews model
