+++ sokoconnect/api/views.py
@@ -40,7 +40,6 @@ class UsersViewSet(viewsets.ModelViewSet):
     filterset_fields = ['usertype']
 
 
-
 class ProductViewSet(viewsets.ModelViewSet):
     queryset = Product.objects.all()
     serializer_class = ProductSerializer
